Fix grammar in code example
Change line section in the sentence, "the code below is jibberish, but is what method chaining is" in line 17. The dependent clause, "but is what method chaining is" needs to be fixed -- as it is unclear what is being expressed. The word "jibberish" is also misspelt and should be 'gibberish' (you can check the veracity of this in any reputable dictionary). The proposed fix -- "the code below is gibberish, but that is what method chaining is." -- eliminates the lack of clarity.